+++ mobile/src/main/java/ch/epfl/sdp/mobile/application/chess/engine2/core/Color.kt
@@ -1,0 +1,6 @@
+package ch.epfl.sdp.mobile.application.chess.engine2.core
+
+enum class Color {
+  Black,
+  White
+} // TODO : Move this.
