+++ mobile/src/main/java/ch/epfl/sdp/mobile/application/chess/engine2/core/ranks/AttackTowardsRank.kt
@@ -1,0 +1,19 @@
+package ch.epfl.sdp.mobile.application.chess.engine2.core.ranks
+
+import ch.epfl.sdp.mobile.application.chess.engine2.core.*
+
+/**
+ * An implementation of [Rank] which attacks and moves towards a [List] of possible directions.
+ *
+ * @param directions the [List] of possible directions.
+ */
+abstract class AttackTowardsRank(private val directions: List<Delta>) : Rank {
+
+  override fun AttackScope.attacks(color: Color, position: Position) {
+    for (direction in directions) {
+      attackTowards(direction, color, position)
+    }
+  }
+
+  override fun ActionScope.actions(color: Color, position: Position) = likeAttacks(color, position)
+}
