+++ mobile/src/main/java/ch/epfl/sdp/mobile/application/chess/engine2/core/ranks/Bishop.kt
@@ -1,0 +1,6 @@
+package ch.epfl.sdp.mobile.application.chess.engine2.core.ranks
+
+import ch.epfl.sdp.mobile.application.chess.engine2.core.Delta.Directions
+
+/** A rank implementation for bishops. */
+object Bishop : AttackTowardsRank(Directions.Diagonals)
