--- mobile/src/main/java/ch/epfl/sdp/mobile/application/chess/engine2/rank/pieces/Pawn.kt
@@ -1,2 +1,0 @@
-package ch.epfl.sdp.mobile.application.chess.engine2.rank.pieces
-
